fixing a lot. sorry for the loose comment.
